main: move out lsblk to its own module
